+++ docs/versionhistory.rst
@@ -5,6 +5,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- **BACKWARDS INCOMPATIBLE** Switched Sentry reporter to use sentry-sdk instead of raven
 - Added support for Python 3.10
 - Dropped support for Python 3.5 and 3.6